Airflow/IS3107GP20_requirements.txt
# reportlab
# matplotlib
# pendulum==2.1.2

python-dotenv==1.0.1
reportlab==4.3.1
pandas==2.2.3
matplotlib==3.10.1
kaggle==1.7.4.2
pymysql==1.1.1

mysql-connector-python==9.2.0

absl-py
google-cloud-bigquery==3.30.0
google-cloud-storage==3.1.0
db-dtypes==1.4.2
google-cloud-bigquery-storage==2.29.1